Revert "Added Undo command (#6)"
This reverts commit 2ba34dfcdb0d4cf3d7f3584763b193e3cdb62338.
