+++ Cargo.toml
@@ -1,20 +1,20 @@
 [package]
 name = "manga_updater"
-version = "0.6.9"
+version = "0.6.8"
 authors = ["Yvan Sanson <sansonyvan@gmail.com>"]
-edition = "2021"
+edition = "2018"
 resolver = "2"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 scraper = "0.12.0"
-reqwest = { version = "0.11.6" }
-tokio = { version = "1.14.0", features = ["full"] }
-futures = "0.3.18"
-structopt = "0.3.25"
+reqwest = { version = "0.11.3" }
+tokio = { version = "1.6.1", features = ["full"] }
+futures = "0.3.15"
+structopt = "0.3.21"
 csv = "1.1.6"
 colour = "0.6.0"
-text_io = "0.1.9"
-open = "2.0.1"
+text_io = "0.1.8"
+open = "1.7.0"
 serial_test = "0.5.1"
