+++ src/commands/mod.rs
@@ -14,8 +14,7 @@ mod remove;
 mod open;
 /// Unread command logic
 mod unread;
-/// Restore command logic
-mod undo;
+
 
 use std::path::PathBuf;
 use crate::commands::list::list_chapters;
@@ -27,7 +26,6 @@ use crate::commands::import::import_file;
 use crate::commands::remove::remove_manga;
 use crate::commands::open::open_manga;
 use crate::commands::unread::unread_chapter;
-use crate::commands::undo::restore_csv;
 
 /// Lists the different mangas and their possible updates.
 /// Passes the logic to the list mod.
@@ -133,8 +131,4 @@ pub fn unread(from: Option<PathBuf>, url: Option<String>, verbose: bool) {
         None => println!("No number provided. Please provide a line number to reset."),
         Some(line_number) => unread_chapter(from, line_number.as_str(), verbose)
     }
-}
-
-pub fn undo(from: Option<PathBuf>,verbose: bool) {
-    restore_csv(from, verbose)
 }
